+++ src/providers/sorbetContentProvider.ts
@@ -1,16 +1,23 @@
-import { CancellationToken, TextDocumentContentProvider, Uri } from 'vscode';
+import * as vscode from 'vscode';
 import { TextDocumentItem } from 'vscode-languageclient';
-import { SorbetExtensionContext } from './sorbetExtensionContext';
+import { SorbetExtensionContext } from '../sorbetExtensionContext';
 
 /**
  * URI scheme supported by {@link SorbetContentProvider}.
  */
 export const SORBET_SCHEME = 'sorbet';
 
+export function registerSorbetContentProvider(context: SorbetExtensionContext): vscode.Disposable {
+  return vscode.workspace.registerTextDocumentContentProvider(
+    SORBET_SCHEME,
+    new SorbetContentProvider(context),
+  );
+}
+
 /**
  * Content provider for URIs with `sorbet` scheme.
  */
-export class SorbetContentProvider implements TextDocumentContentProvider {
+export class SorbetContentProvider implements vscode.TextDocumentContentProvider {
   private readonly context: SorbetExtensionContext;
 
   constructor(context: SorbetExtensionContext) {
@@ -21,8 +28,8 @@ export class SorbetContentProvider implements TextDocumentContentProvider {
    * Provide textual content for a given uri.
    */
   public async provideTextDocumentContent(
-    uri: Uri,
-    token?: CancellationToken,
+    uri: vscode.Uri,
+    token?: vscode.CancellationToken,
   ): Promise<string> {
     let content: string;
     const { activeLanguageClient: client } = this.context.statusProvider;
@@ -30,9 +37,7 @@ export class SorbetContentProvider implements TextDocumentContentProvider {
       this.context.log.info('ContentProvider: Retrieving file contents', uri);
       const response = await client.sendRequest<TextDocumentItem>(
         'sorbet/readFile',
-        {
-          uri: uri.toString(),
-        },
+        { uri: uri.toString() },
         token,
       );
       content = response?.text ?? '';